+++ lib/smartrefresh.dart
@@ -1,2 +1,2 @@
-export 'src/pullToRefresh.dart';
-export 'src/refreshController.dart';
+export 'src/pull_to_refresh.dart';
+export 'src/refresh_controller.dart';